Utilisation de size_t au lieu de int.

+++ include/fundamentals.h
@@ -91,7 +91,7 @@ template<typename S,typename T> class action {
 public:
     action() {}
     virtual ~action() {}
-    virtual void perform(const S* string,int string_length,int remaining_distance,T data)=0;
+    virtual void perform(const S* string,size_t string_length,int remaining_distance,T data)=0;
     virtual T result()=0;
 };
 
@@ -99,7 +99,7 @@ template<typename S,typename T> class filter {
 public:
     filter() {}
     virtual ~filter() {}
-    virtual T perform(const S* string,int string_length,int remaining_distance,T data)=0;
+    virtual T perform(const S* string,size_t string_length,int remaining_distance,T data)=0;
 };
 
 template<typename T> class null_reader_writer {
